Consolidate climb hold + zone search into a single board form
Replaces the separate "Holds" and "Zone" accordion sections with one
"Board" section that shares a single board image. Wraps the whole stack
(BoardRenderer + filter overlay + heatmap overlay + zone overlay) in
ZoomableBoard so pinch-zoom, ctrl+wheel, and drag-to-pan now work for the
search board, matching the climb-view experience.

The combined form keeps both interactions on the same board: tap holds to
filter via the popover, drag the zone rect's corners (or its invisible
thick grab-stroke) to constrain results. Zone enable now defaults to the
full board so it doesn't surprise-clear existing hold filters; the user
shrinks by dragging inward.

When the user commits a zone change, any selected hold whose centre falls
outside the new rectangle is silently dropped from holdsFilter in the
same updateFilters call, keeping the two filters consistent.
