+++ packages/web/e2e/help-screenshots.spec.ts
@@ -56,15 +56,15 @@ test.describe('Help Page Screenshots', () => {
   test('search by hold', async ({ page }) => {
     await page.getByRole('button', { name: 'Open filters' }).click();
     await page.locator('[data-swipeable-drawer="true"]:visible').first().waitFor({ timeout: 10_000 });
-    // Expand the "Holds" accordion section.
-    await page.getByText('Holds', { exact: true }).click();
+    // Expand the combined Board accordion section (holds + zone).
+    await page.getByText('Board', { exact: true }).click();
     await page.screenshot({ path: `${SCREENSHOT_DIR}/search-by-hold.png` });
   });
 
   test('heatmap', async ({ page }) => {
     await page.getByRole('button', { name: 'Open filters' }).click();
     await page.locator('[data-swipeable-drawer="true"]:visible').first().waitFor({ timeout: 10_000 });
-    await page.getByText('Holds', { exact: true }).click();
+    await page.getByText('Board', { exact: true }).click();
     await page.getByRole('button', { name: 'Show Heatmap' }).click();
     await page.waitForSelector('text=Loading heatmap...', { state: 'hidden', timeout: 10_000 }).catch(() => {});
     await page.waitForSelector('button:has-text("Hide Heatmap")', { state: 'visible' });
